+++ src/main/java/de/maxhenkel/car/integration/theoneprobe/TileInfoProvider.java
@@ -8,6 +8,7 @@ import mcjty.theoneprobe.api.IProbeHitData;
 import mcjty.theoneprobe.api.IProbeInfo;
 import mcjty.theoneprobe.api.IProbeInfoProvider;
 import mcjty.theoneprobe.api.ProbeMode;
+import mcjty.theoneprobe.apiimpl.styles.ProgressStyle;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Player;
@@ -38,15 +39,13 @@ public class TileInfoProvider implements IProbeInfoProvider {
         } else if (te instanceof TileEntityGenerator generator) {
             FluidStack fluid = generator.getFluidInTank(0);
             if (!fluid.isEmpty()) {
-                // TODO Add back when The One Probe updated
-                // iProbeInfo.tankHandler(generator, new ProgressStyle().suffix(" mb"));
+                iProbeInfo.tankHandler(generator, new ProgressStyle().suffix(" mb"));
             }
         } else if (te instanceof TileEntityTank tank) {
             FluidStack fluid = tank.getFluidInTank(0);
             if (!fluid.isEmpty()) {
                 iProbeInfo.text(fluid.getHoverName());
-                // TODO Add back when The One Probe updated
-                // iProbeInfo.tankHandler(tank, new ProgressStyle().suffix(" mb"));
+                iProbeInfo.tankHandler(tank, new ProgressStyle().suffix(" mb"));
             }
         }
     }