+++ src/main/java/de/maxhenkel/car/integration/waila/HUDHandlerTank.java
@@ -20,7 +20,7 @@ public class HUDHandlerTank implements IBlockComponentProvider, IServerDataProvi
 
     @Override
     public void appendTooltip(ITooltip iTooltip, BlockAccessor blockAccessor, IPluginConfig iPluginConfig) {
-        FluidStack stack = FluidStack.parseOptional(blockAccessor.getBlockEntity().getLevel().registryAccess()/*TODO*/, blockAccessor.getServerData().getCompound("fluid"));
+        FluidStack stack = FluidStack.parseOptional(blockAccessor.getBlockEntity().getLevel().registryAccess(), blockAccessor.getServerData().getCompound("fluid"));
 
         if (stack.isEmpty()) {
             iTooltip.add(Component.translatable("tooltip.waila.tank.no_fluid"));
@@ -33,7 +33,7 @@ public class HUDHandlerTank implements IBlockComponentProvider, IServerDataProvi
     @Override
     public void appendServerData(CompoundTag compoundTag, BlockAccessor blockAccessor) {
         TileEntityTank tank = (TileEntityTank) blockAccessor.getBlockEntity();
-        compoundTag.put("fluid", tank.getFluid().save(blockAccessor.getBlockEntity().getLevel().registryAccess()/*TODO*/));
+        compoundTag.put("fluid", tank.getFluid().save(blockAccessor.getBlockEntity().getLevel().registryAccess()));
     }
 
     @Override